libbladeRF, bladeRF-cli: Reworked cal data generation and misc cleanup
The function to create cal data now is a wrapper around
bladerf_alloc_image() which initializes the returned image with
calibration data.

The associated CLI command now allows this to be either written to a
file or to the device directly.

API header cleanup: Pushed the lower level functions to to the end of
the header, so the functions people will use more frequently will be at
the top of the generated documentation.

Also added a missing API_EXPORT and made the Doxygen color scheme a little
more blue. (Tired of grayscale ;) )

Windows cleanup: Added get_timestamp() implementation, cleaned up
warnings, address some portability issues.

+++ host/libraries/libbladeRF/src/file_ops.c
@@ -63,7 +63,7 @@ int file_read(FILE *f, char *buf, size_t len)
 
 ssize_t file_size(FILE *f)
 {
-    int rv = BLADERF_ERR_IO;
+    ssize_t rv = BLADERF_ERR_IO;
     long int fpos = ftell(f);
     ssize_t len;
